This optimization based on article  
[Cassio Neri, Lorenz Schneiderhttps - Euclidean Affine Functions and Applications to Calendar Algorithms - 2021](https://arxiv.org/pdf/2102.06959.pdf)

It avoids loops and minimizes branching in calendar calculations.

Benchmarks:
```
|                                                     Method |      Mean | Error |       Min |       Max | Allocated |
|----------------------------------------------------------- |----------:|------:|----------:|----------:|----------:|
|           'DateTime.GetDate(out year, out month, out day)' |  7.697 ns |    NA |  7.697 ns |  7.697 ns |         - |
| 'Optimized DateTime.GetDate(out year, out month, out day)' |  5.114 ns |    NA |  5.114 ns |  5.114 ns |         - |


|                     Method |     Mean | Error |      Min |      Max | Allocated |
|--------------------------- |---------:|------:|---------:|---------:|----------:|
|              DateTime.Year | 4.289 ns |    NA | 4.289 ns | 4.289 ns |         - |
|             DateTime.Month | 6.534 ns |    NA | 6.534 ns | 6.534 ns |         - |
|               DateTime.Day | 6.946 ns |    NA | 6.946 ns | 6.946 ns |         - |
|  'Optimized DateTime.Year' | 2.945 ns |    NA | 2.945 ns | 2.945 ns |         - |
| 'Optimized DateTime.Month' | 3.161 ns |    NA | 3.161 ns | 3.161 ns |         - |
|   'Optimized DateTime.Day' | 3.638 ns |    NA | 3.638 ns | 3.638 ns |         - |

```

It is difficult to explain values of magic constants of the algorithm in comments.
They are inferred algebraically.

* Get rid of `GetDatePart()` function and inline its functionality directly into `Year, Month, Day, DayOfYear` properties