@danmoseley 2022-07-24
Do we have test coverage here sufficient to give high confidence? If you can think of any other useful test cases this PR is a good time to add them.