@danmoseley 2022-07-24
Maybe @lemire is interested in this trick.