@SergeiPavlov 2022-07-24
> Do we have test coverage here sufficient to give high confidence? If you can think of any other useful test cases this PR is a good time to add them.

I tested it for every possible date (range 1/1/0001 .. 12/31/9999). Results are identical to current implementation.
Of course it does not replace testing in dotnet workflow.