@lemire 2022-07-24
@danmoseley I knew of the article, but I am happy to see this work come to dotnet.