@danmoseley 2022-07-28
I've compared the paper to the code changes here and I can reproduce the calculations, so the code makes sense to me.  (Although, I see `r2 = u2 % 2^32 / 2939745 / 4`, is turning into `(ushort)((uint)u2 / 11758980)` .. clearly produces the right result, but I would think precedence would run the divisions one after the other, rather than multiplying .. what am I missing about their notation..)

I added some more random tests locally and they all pass, which should not be a surprise since you say you've tried all possible valid dates yourself.

> Adding [MethodImpl(MethodImplOptions.AggressiveInlining)] to GetDatePart() makes properties Year, Month, Day, DayOfYear significantly more efficient. The compiler discards unrelated branches of GetDatePart().

Would it make for clearer code, albeit with a bit more duplication, to manually inline the appropriate parts inside each of these? Then it will be self evident that the unrelated branches are discarded.

Re my question about perf tests, I guess this covers it indirectly. Did you run this test before/after? But, it would probably be good to add benchmarks in this file for Year, Month, Day, DayOfYear individually, especially if you end up copy/pasting instead of inlining as suggested above.
https://github.com/dotnet/performance/blob/e60e729d0008d8522b518b11c06b9803f6d674ac/src/benchmarks/micro/libraries/System.Runtime/Perf.DateTime.cs#L34