@danmoseley 2022-07-28
@tannergooding in case he is interested in the math..