@danmoseley 2022-07-28
I notice we are missing tests for DayOfYear that use randomized inputs. Year, Month, Day are indirectly covered by the ToString() tests with randomized inputs
https://github.com/dotnet/runtime/blob/cc7ccfd0763b1f3dca173268ac7295ab4afeefb8/src/libraries/System.Runtime/tests/System/DateTimeTests.cs#L2149
Those ToString() tests indirectly test Year, Month, Day getters.