@danmoseley 2022-07-28
@richlander when we're using an algorithm from a paper like this, ie., there's no particular software license, I assume credit in a comment is fine, and no entry in THIRD-PARTY-NOTICES.TXT is expected?