@SergeiPavlov 2022-07-28
> Would it make for clearer code, albeit with a bit more duplication, to manually inline the appropriate parts inside each of these? Then it will be self evident that the unrelated branches are discarded.

Inlined them