@SergeiPavlov 2022-07-28
> Re my question about perf tests, I guess this covers it indirectly. Did you run this test before/after? But, it would probably be good to add benchmarks in this file for Year, Month, Day, DayOfYear individually, especially if you end up copy/pasting instead of inlining as suggested above. (If you add benchmarks, that would of course be a PR into dotnet/performance that we'd merge after code flows, but you could run them locally and report here.)

Locally I have a source file with both: original & new implementations. The first benchmarking result were before applying `[MethodImpl(MethodImplOptions.AggressiveInlining)]`.

Added individual benchmarks for the properties. Results ar in the PR description. The most significant gain (50%) is for `Month` which was most heavy because of loop.