@danmoseley 2022-07-28
@richlander is going to check on licensing here. Marking **do not merge** meantime. The change is otherwise fine.