@richlander 2022-07-28
Right. I sent mail to the authors of the paper + @SergeiPavlov.