@cassioneri 2022-07-31
Lorenz and I are glad to see that our algorithm might go into .NET. We wish to thank, especially, Sergei for implementing and creating the PR.

On the licensing question sent by email, Lorenz and I have discussed the matter and we agreed with your second bullet point which I quote:

> Agree that the code was primarily informed by the paper and that it is licensed permissively such that we can land the PR.

It is clear that Sergei did implement the algorithm as explained in the paper. In terms of implementation, we consider his C# code substantially different from the GPL-licensed C++ code found in my repository.

For the reasons above, as far as Lorenz and I are concerned, we do not believe that GPL applies to Sergei's PR. As I said above, we would like to see the PR merged into the .NET's main branch for the benefit of .NET users. We would appreciate (but not require) if the release notes could mention our work.