@richlander 2022-07-31
We will add in release notes and a blog post (including the release blog post). Thanks much!

@JonDouglas