@IDisposable 2022-08-01
> I've compared the paper to the code changes here and I can reproduce the calculations, so the code makes sense to me. (Although, I see `r2 = u2 % 2^32 / 2939745 / 4`, is turning into `(ushort)((uint)u2 / 11758980)` .. clearly produces the right result, but I would think precedence would run the divisions one after the other, rather than multiplying .. what am I missing about their notation..)

Ignoring the `% 2^32` part... the basic math rule at hand here is that `X / Y / Z === X / (Y * Z)`. The easiest way to grok this is an example:  64 / 4 / 4 -> left-to-right apply first divisions `(64 / 4) / 4 == 16 / 4` -> apply remaining division `16 / 4 == 4`... now let's pre-apply the second division (which can be thought of as _scaling up_ the divisor of the first one... so `64 / ( 4 * 4) == 64 / 16` and then apply the first division so `64 / 16 == 4`... same answer... and if we KNOW the right-side divisors are constant there's no reason not to pre-multiply.