@SergeiPavlov 2022-08-02
> `((X / Y) / Z) * Z = a * Z` --> move over the `Z` by multiplying both sides by `Z`
> `(X / Y) = a * Z` --> which isolates `X / Y` on left

Hm, this implication does not work for integer division in general case.
Counter-example:
```
X = 21
Y = 3
Z = 2
a = X / Y / Z = 3
```
`((21 / 3) / 2) * 2` = `3 * 2`
but
`(21 / 3)` ≠ `3 * 2`