@IDisposable 2022-08-02
Certainly there are truncation errors for integers where significance is lost, that's why I was talking about it in algebraic form :)