@SergeiPavlov 2022-08-02
The proof for integer division:

let   X = YA+B;     X/Y = A = ZQ+R
where Y = (B+M), 1 <= M
      Z = (R+N), 1 <= N

then:
YZ = (B+M)(R+N) = BR+BN +MR+MN

X = YA + B = Y(ZQ+R) + B = YZQ + YR + B = (YZ)Q + (B+M)R + B = (YZ)Q + BR + MR + B

X/(YZ) = ((YZ)Q + BR + MR + B) / YZ 
 = Q + (BR+MR+B)/YZ
 = Q + (BR+MR+B)/(BR+BN+RM+MN) = Q       because  BR+MR+B < BR+BN+RM+MN always