@SergeiPavlov 2022-08-02
> `((Y * (X / Y)) / Z = a * Y` --> move over the Y by multiplying both sides by Y
> `(X / Z) = a * Y` --> which isolates X / Z on left

again `(Y * (X / Y)`  is not always the same as `X`
Rational numbers algebra rules are not applicable for integer arithmetics.